+++ sys-kernel/git-sources/git-sources-3.6.11.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2013 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: sys-kernel/git-sources/git-sources-3.6.11.ebuild,v 1.5 2013/03/19 21:20:43 -tclover Exp $
+# $Header: sys-kernel/git-sources/git-sources-3.6.11.ebuild,v 1.5 2013/06/06 16:32:42 -tclover Exp $
 
 EAPI=5
 
@@ -101,7 +101,9 @@ src_prepare() {
 			epatch "${WORKDIR}"/patches/preempt-desktop-tune.patch
 		fi
 	fi
-	use bfq && epatch "${FILESDIR}"/${bfq_src}
+	if use bfq; then
+		bzip2 -cd "${FILESDIR}"/${bfq_src} | patch -p1 || die
+	fi
 	use bld && epatch "${DISTDIR}"/${bld_src}
 	use uksm && epatch "${DISTDIR}"/${uksm_src}
 	rm -fr .git* b
